requirements.txt
appdirs==1.4.4
astor==0.8.1
asttokens==3.0.0
beautifulsoup4==4.13.5
blinker==1.9.0
Brotli==1.1.0
cairocffi==1.7.1
CairoSVG==2.8.2
certifi==2025.8.3
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.3
choreographer==1.0.10
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cssselect2==0.8.0
cycler==0.12.1
dash==2.18.2
dash-bootstrap-components==1.7.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dash_daq==0.5.0
decorator==5.2.1
defusedxml==0.7.1
dtale==3.18.2
et_xmlfile==2.0.0
executing==2.2.0
Flask==3.0.3
Flask-Compress==1.18
fonttools==4.59.0
future==1.0.0
git-filter-repo==2.47.0
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
ipython==9.4.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
jsonpickle==4.1.1
kaleido==1.0.0
kiwisolver==1.4.8
logistro==1.1.0
lz4==4.4.4
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
-e git+https://github.com/coreyprator/Ableton-midi-bench.git@ad706292154c7c75bd7c998b6df788e1290dba1e#egg=midi_benchmark
mido==1.3.3
missingno==0.5.2
more-itertools==10.7.0
music21==9.7.1
narwhals==2.1.2
nest-asyncio==1.6.0
networkx==3.5
numpy==2.2.6
opencv-python==4.10.0.84
openpyxl==3.1.5
orjson==3.11.2
packaging==25.0
pandas==2.3.1
pandasgui==0.2.15
parso==0.8.4
patsy==1.0.1
pillow==10.4.0
pkginfo==1.12.1.2
plotly==6.3.0
pluggy==1.6.0
pretty_midi==0.2.10
prompt_toolkit==3.0.51
pure_eval==0.2.3
pyarrow==21.0.0
pycparser==2.22
Pygments==2.19.2
pynput==1.8.1
pyparsing==3.2.3
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
PyQtWebEngine==5.15.7
PyQtWebEngine-Qt5==5.15.2
pytest==8.4.1
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==311
pyzstd==0.17.0
qtstylish==0.1.5
requests==2.32.5
retrying==1.4.2
scikit-learn==1.7.1
scipy==1.15.3
seaborn==0.13.2
setuptools==80.9.0
shapely==2.1.1
simplejson==3.20.1
six==1.17.0
soupsieve==2.7
squarify==0.4.4
stack-data==0.6.3
statsmodels==0.14.5
strsimpy==0.2.1
tabulate==0.9.0
threadpoolctl==3.6.0
tinycss2==1.4.0
traitlets==5.14.3
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
Werkzeug==3.0.6
wheel==0.45.1
wordcloud==1.9.4
xarray==2025.8.0
xlrd==2.0.2
zipp==3.23.0
